Adding int,short,char types in C++ and int in Scala; changing multiplicity type and date type to int; extra checks in type checker

+++ ddbtoaster/srccpp/lib/event.hpp
@@ -14,7 +14,7 @@
 
 namespace dbtoaster {
 
-typedef long date;
+typedef int date;
 
 /**
  * Type definitions of data-structures used for representing events.
